+++ src/app/game/game.py
@@ -1,5 +1,6 @@
 from flask import render_template, url_for, redirect, Blueprint
 
+from src.app.model.board.threeboard import ThreeBoard
 from src.app.model.mood import Mood
 from src.app.model.status import Status
 from src.version.version import __version__
@@ -25,7 +26,10 @@ def construct_blueprint(redis, messages):
         else:
             player_two_active = True
 
-        game_state = get_game_state(redis)
+        game_mode = redis.get("gameMode")
+        board = redis.get_complex("board")
+        game_state = get_game_state(redis, board)
+
         print(game_state)
         print("user id")
         print(user_id)
@@ -38,34 +42,37 @@ def construct_blueprint(redis, messages):
             if game_state == Status.DRAW:
                 notification_header = messages.load("game.end.draw.header")
                 notification_message = messages.load("game.end.draw.1.message")
-                notification_icon = messages.load("game.end.draw.1.icon")
+                notification_icon = messages.load("game.end.draw.icon")
 
             elif game_state == Status.PLAYER_ONE_WINS:
                 if redis.get("player1") == user_id:
                     notification_header = messages.load_with_params("game.end.win.header", [redis.get("player1")])
+                    notification_icon = messages.load("game.end.win.icon")
                     notification_message = messages.load("game.end.win.1.message")
-                    notification_icon = messages.load("game.end.win.1.icon")
                     notification_mood = Mood.HAPPY.value
                 else:
                     notification_header = messages.load_with_params("game.end.lose.header", [redis.get("player2")])
+                    notification_icon = messages.load("game.end.lose.icon")
                     notification_message = messages.load("game.end.lose.3.message")
-                    notification_icon = messages.load("game.end.lose.3.icon")
                     notification_mood = Mood.SAD.value
 
             elif game_state == Status.PLAYER_TWO_WINS:
                 if redis.get("player2") == user_id:
                     notification_header = messages.load_with_params("game.end.win.header", [redis.get("player2")])
+                    notification_icon = messages.load("game.end.win.icon")
                     notification_message = messages.load("game.end.win.3.message")
-                    notification_icon = messages.load("game.end.win.3.icon")
                     notification_mood = Mood.HAPPY.value
                 else:
                     notification_header = messages.load_with_params("game.end.lose.header", [redis.get("player1")])
+                    notification_icon = messages.load("game.end.lose.icon")
                     notification_message = messages.load("game.end.lose.1.message")
-                    notification_icon = messages.load("game.end.lose.1.icon")
                     notification_mood = Mood.SAD.value
 
         return render_template(
             "game.html",
+            board=board,
+            playableSquare=redis.get("playableSquare"),
+            innerStates=redis.get_complex("innerStates"),
             zero=redis.get("0"),
             one=redis.get("1"),
             two=redis.get("2"),
@@ -77,6 +84,7 @@ def construct_blueprint(redis, messages):
             eight=redis.get("8"),
             gameComplete=game_complete,
             gameId=game_id,
+            gameMode=game_mode,
             notificationActive=notification_active,
             notificationHeader=notification_header,
             notificationMessage=notification_message,
@@ -93,10 +101,12 @@ def construct_blueprint(redis, messages):
         )
 
     @game_page.route("/game/<game_id>/place-move/<user_id>/<square>")
-    def place_move(game_id, user_id, square):
+    def place_standard_move(game_id, user_id, square):
         # Set player's move
         symbol = redis.get(user_id)
-        redis.set(square, symbol)
+        board_list = redis.get_complex("board")
+        board_list[int(square)] = int(symbol)
+        redis.set_complex("board", board_list)
 
         # Switch player turn
         if redis.get("whoseTurn") == 'player1':
@@ -106,16 +116,39 @@ def construct_blueprint(redis, messages):
 
         return redirect(url_for("game_page.game", game_id=game_id, user_id=user_id))
 
+    @game_page.route("/game/<game_id>/place-move/<user_id>/<outer_square>/<inner_square>")
+    def place_ultimate_move(game_id, user_id, outer_square, inner_square):
+        # Set player's move
+        symbol = redis.get(user_id)
+        board = redis.get_complex("board")
+        board[int(outer_square)][int(inner_square)] = int(symbol)
+        redis.set_complex("board", board)
+
+        print("[place_ultimate_move] outer_square: " + outer_square)
+        print("[place_ultimate_move] inner_square: " + inner_square)
+        print(board)
+
+        # Set next playable outer square
+        if get_game_state(redis, board[int(inner_square)]) != Status.IN_PROGRESS:
+            redis.set("playableSquare", "-1")  # -1 is all squares...
+        else:
+            redis.set("playableSquare", inner_square)
+
+        print("[place_ultimate_move] playableSquare: " + redis.get("playableSquare"))
+
+        # Switch player turn
+        if redis.get("whoseTurn") == 'player1':
+            redis.set("whoseTurn", "player2")
+        elif redis.get("whoseTurn") == 'player2':
+            redis.set("whoseTurn", "player1")
+
+        return redirect(url_for("game_page.game", game_id=game_id, user_id=user_id))
+
+    # Closing return
     return game_page
 
 
-def get_game_state(redis):
-    board = {
-        "0": redis.get("0"), "1": redis.get("1"), "2": redis.get("2"),
-        "3": redis.get("3"), "4": redis.get("4"), "5": redis.get("5"),
-        "6": redis.get("6"), "7": redis.get("7"), "8": redis.get("8"),
-    }
-
+def get_game_state(redis, board):
     winning_combos = [
         [0, 1, 2],
         [3, 4, 5],
@@ -127,22 +160,35 @@ def get_game_state(redis):
         [2, 4, 6]
     ]
 
-    print(board)
+    if isinstance(board[0], list):
+        print("[get_game_state] board[0]: " + str(board[0]))
+        outer_state = Status.IN_PROGRESS
+        inner_states = []
+        for outer_square in board:
+            inner_state = get_game_state(redis, outer_square)
+            inner_states.append(inner_state.value)
+            print("[get_game_state] inner_states: " + str(inner_states))
+            if inner_states.count('1') == 0:
+                return Status.DRAW
+        redis.set_complex("innerStates", inner_states)
+        return get_game_state(redis, create_false_board(inner_states))
 
-    if list(board.values()).count("0") == 0:
+    if board.count(0) == 0:
         return Status.DRAW
+        # FixMe :: this check needs to happen after test each player has won...
+        # Note :: There is probs a clean way to split this up so that you don't iterate when not nec...
 
-    print("count: " + str(list(board.values()).count("1")))
-    if (list(board.values()).count("1")) >= 3:
-        player_moves = get_player_moves("1", board)
+    print("count: " + str(board.count(1)))
+    if (board.count(1)) >= 3:
+        player_moves = get_player_moves(1, board)
         print(player_moves)
         for combo in winning_combos:
             print(combo)
             if set(combo).issubset(set(player_moves)):
                 return Status.PLAYER_ONE_WINS
 
-    if (list(board.values()).count("2")) >= 3:
-        player_moves = get_player_moves("2", board)
+    if (board.count(2)) >= 3:
+        player_moves = get_player_moves(2, board)
         for combo in winning_combos:
             if set(combo).issubset(set(player_moves)):
                 return Status.PLAYER_TWO_WINS
@@ -151,4 +197,32 @@ def get_game_state(redis):
 
 
 def get_player_moves(player, board):
-    return [int(k) for k, v in board.items() if v == player]
+    player_moves = []
+    for index in range(len(board)):
+        if board[index] == player:
+            player_moves.append(index)
+
+    return player_moves
+
+
+def convert_states_to_symbols(state):
+    if state == 1: return 0
+    if state == 2: return 0
+    if state == 3: return 1
+    if state == 4: return 2
+
+
+def create_false_board(states):
+    board = ThreeBoard()
+    board.top_lhs = convert_states_to_symbols(states[0])
+    board.top_mid = convert_states_to_symbols(states[1])
+    board.top_rhs = convert_states_to_symbols(states[2])
+    board.mid_lhs = convert_states_to_symbols(states[3])
+    board.mid_mid = convert_states_to_symbols(states[4])
+    board.mid_rhs = convert_states_to_symbols(states[5])
+    board.bot_lhs = convert_states_to_symbols(states[6])
+    board.bot_mid = convert_states_to_symbols(states[7])
+    board.bot_rhs = convert_states_to_symbols(states[8])
+
+    print("[create_false_board] board: " + str(board.list()))
+    return board.list()